doc_service/requirements.txt
fastapi
uvicorn
python-dotenv
docx2txt
PyPDF2
markdown
